feat: add auto termination enum
