--- embrace-test-fakes/src/main/kotlin/io/embrace/android/embracesdk/fakes/FakeTracingApi.kt
@@ -1,46 +1,0 @@
-package io.embrace.android.embracesdk.fakes
-
-import io.embrace.android.embracesdk.spans.EmbraceSpan
-import io.embrace.android.embracesdk.spans.EmbraceSpanEvent
-import io.embrace.android.embracesdk.spans.ErrorCode
-import io.embrace.android.embracesdk.spans.TracingApi
-
-class FakeTracingApi : TracingApi {
-
-    val createdSpans: MutableList<String> = mutableListOf()
-
-    override fun createSpan(name: String, parent: EmbraceSpan?): EmbraceSpan? {
-        createdSpans.add(name)
-        return null
-    }
-
-    override fun startSpan(name: String, parent: EmbraceSpan?, startTimeMs: Long?): EmbraceSpan? {
-        TODO("Not yet implemented")
-    }
-
-    override fun <T> recordSpan(
-        name: String,
-        parent: EmbraceSpan?,
-        attributes: Map<String, String>?,
-        events: List<EmbraceSpanEvent>?,
-        code: () -> T,
-    ): T {
-        TODO("Not yet implemented")
-    }
-
-    override fun recordCompletedSpan(
-        name: String,
-        startTimeMs: Long,
-        endTimeMs: Long,
-        errorCode: ErrorCode?,
-        parent: EmbraceSpan?,
-        attributes: Map<String, String>?,
-        events: List<EmbraceSpanEvent>?,
-    ): Boolean {
-        TODO("Not yet implemented")
-    }
-
-    override fun getSpan(spanId: String): EmbraceSpan? {
-        TODO("Not yet implemented")
-    }
-}
